Split misc routines from math.cc
