+++ chainerx_cc/examples/mnist/train_mnist.cc
@@ -20,7 +20,7 @@
 #include "chainerx/routines/creation.h"
 #include "chainerx/routines/linalg.h"
 #include "chainerx/routines/manipulation.h"
-#include "chainerx/routines/math.h"
+#include "chainerx/routines/misc.h"
 #include "chainerx/routines/reduction.h"
 #include "chainerx/shape.h"
 #include "chainerx/slice.h"
